requirements.txt
Django
django-crispy-forms
mysqlclient
Pillow
pkg-resources
pytz
sqlparse
djangorestframework
